optimize leaderboard

+++ backend/src/controllers/expenseController.js
@@ -1,4 +1,6 @@
 const Expense = require('../models/Expense');
+const User = require('../models/User');
+const { Sequelize } = require('sequelize');
 
 exports.createExpense = async (req, res) => {
     try {
@@ -8,6 +10,10 @@ exports.createExpense = async (req, res) => {
             amount, description, category,
             userId: req.user.id
         });
+        const user = await User.findByPk(req.user.id);
+        user.totalExpense = (parseFloat(user.totalExpense) || 0) + parseFloat(amount);
+        await user.save();
+
         res.status(201).json(newExpense);
     } catch (error) {
         console.log(Error, error);
@@ -29,15 +35,23 @@ exports.getExpenses = async (req, res) => {
 exports.deleteExpense = async (req, res) => {
     try {
         const { id } = req.params;
-        const deleteExpense = await Expense.destroy({
+        const expense = await Expense.findOne({
             where: { id },
             userId: req.user.id
         });
-        if (deleteExpense) {
-            res.status(204).send();
-        } else {
-            res.status(404).json({ error: 'Expense not found' });
+        if (!expense) {
+            return res.status(404).json({ error: 'Expense not found' });
         }
+
+        const user = await User.findByPk(req.user.id);
+        user.totalExpense = parseFloat(user.totalExpense) - parseFloat(expense.amount);
+        await user.save();
+
+        await Expense.destroy({
+            where: { id }
+        });
+
+        res.status(204).send();        
     } catch (error) {
         res.status(500).json({ error: 'Failed to delete expense' });
     }
